fix(scatac): optional --feature (#3) + stop uint16 overflow in pseudobulk/obs sums (#6, #11)
Three audit fixes in the Python scATAC layer. Counts stay bit-identical on the
GM12878 golden baseline (layers ref/alt/other + X, max_abs_diff=0); cargo test 109 passed.

- #3 (run_counting_sc.py): replace the unconditional `assert feature_file is not None`
  with an optional-feature path — when None, skip the peak intersection and count over
  ALL variants (intersect_file already defaults to vcf_bed). Fixes a crash on the
  documented no-feature default. Validated: no-feature chr20 run completes (49962
  variants vs 2702 peak-filtered), no uns['feature'].
- #6 (as_analysis_sc.py): drop dtype=np.uint16 from the 4 pseudobulk .sum(axis=1) calls.
  scipy get_sum_dtype widens uint16->uint64, so per-peak ref/alt sums no longer wrap at
  >65535 (which silently corrupted the beta-binomial AI inputs). Value-neutral on GM12878
  (max sums ref=3416/alt=1918 << 65535); synthetic 70000: old wraps to 4464, new is 70000.
- #11 (count_alleles_sc.py): same dtype fix on the obs ref_count/alt_count row sums
  (values identical on GM12878, dtype widens uint16->uint64).

No Rust change. #2 (AnnData obs/var orientation) deliberately deferred as a data-contract
decision (separate migration; would transpose the golden + figures + analyzer).

+++ src/analysis/as_analysis_sc.py
@@ -85,8 +85,8 @@ def get_imbalance_sc(
     # region_cutoff = min_count + (2*pseudocount)
     snp_cutoff = 2 * pseudocount
 
-    ref_counts = adata.layers["ref"].sum(axis=1, dtype=np.uint16).T.A1 + pseudocount
-    alt_counts = adata.layers["alt"].sum(axis=1, dtype=np.uint16).T.A1 + pseudocount
+    ref_counts = adata.layers["ref"].sum(axis=1).T.A1 + pseudocount
+    alt_counts = adata.layers["alt"].sum(axis=1).T.A1 + pseudocount
     n_counts = ref_counts + alt_counts
 
     # Calculate dispersion across dataset
@@ -116,8 +116,8 @@ def get_imbalance_sc(
         adata_sub = adata[:, adata.var["group"] == group_name]
 
         # Create count data per group
-        ref_counts_group = adata_sub.layers["ref"].sum(axis=1, dtype=np.uint16).T.A1 + pseudocount
-        alt_counts_group = adata_sub.layers["alt"].sum(axis=1, dtype=np.uint16).T.A1 + pseudocount
+        ref_counts_group = adata_sub.layers["ref"].sum(axis=1).T.A1 + pseudocount
+        alt_counts_group = adata_sub.layers["alt"].sum(axis=1).T.A1 + pseudocount
         n_counts_group = ref_counts_group + alt_counts_group
 
         nonzero_idx = np.where(n_counts_group > snp_cutoff)  # Get indices where counts were found
